+++ pyasn1/codec/ber/decoder.py
@@ -1450,6 +1450,10 @@ TYPE_MAP = {
     univ.Any.typeId: AnyPayloadDecoder()
 }
 
+# deprecated aliases, https://github.com/pyasn1/pyasn1/issues/9
+tagMap = TAG_MAP
+typeMap = TYPE_MAP
+
 # Put in non-ambiguous types for faster codec lookup
 for typeDecoder in TAG_MAP.values():
     if typeDecoder.protoComponent is not None: