+++ pyasn1/codec/ber/encoder.py
@@ -773,6 +773,10 @@ TYPE_MAP = {
     useful.UTCTime.typeId: OctetStringEncoder()
 }
 
+# deprecated aliases, https://github.com/pyasn1/pyasn1/issues/9
+tagMap = TAG_MAP
+typeMap = TYPE_MAP
+
 
 class SingleItemEncoder(object):
     fixedDefLengthMode = None