+++ pyasn1/codec/der/encoder.py
@@ -57,6 +57,10 @@ TYPE_MAP.update({
     univ.Set.typeId: SetEncoder()
 })
 
+# deprecated aliases, https://github.com/pyasn1/pyasn1/issues/9
+tagMap = TAG_MAP
+typeMap = TYPE_MAP
+
 
 class SingleItemEncoder(encoder.SingleItemEncoder):
     fixedDefLengthMode = True