+++ pyasn1/codec/native/decoder.py
@@ -129,6 +129,10 @@ TYPE_MAP = {
     useful.UTCTime.typeId: AbstractScalarPayloadDecoder()
 }
 
+# deprecated aliases, https://github.com/pyasn1/pyasn1/issues/9
+tagMap = TAG_MAP
+typeMap = TYPE_MAP
+
 
 class SingleItemDecoder(object):